# [MM-29241] Image logic refactor