… app/imaging/decode_bench_test.go
… app/imaging/decode_test.go
… app/imaging/encode_test.go
… app/imaging/preview.go
… app/imaging/svg.go
… app/imaging/svg_test.go
… app/imaging/utils.go
… app/imaging/utils_bench_test.go
… app/imaging/utils_test.go
… app/server.go
… app/slack.go
… app/team.go
… app/upload.go
… app/user.go
… i18n/en.json
… model/file_info.go
… services/slackimport/slackimport.go
… tests/fill_test_16bit_rgb.png
… tests/fill_test_16bit_rgb_out.png
… tests/fill_test_16bit_rgba.png
… tests/fill_test_16bit_rgba_out.png
… tests/fill_test_8bit_palette.png
… tests/fill_test_8bit_palette_out.png
… tests/fill_test_8bit_rgb.png
… tests/fill_test_8bit_rgb_out.png
… tests/fill_test_8bit_rgba.png
… tests/fill_test_8bit_rgba_out.png
… tests/fill_test_opaque.png