#### Summary

PR moves most of the image processing logic into a separate `imaging` package under `app`.
Other noticeable changes include:

- Provided implementations for global image decoders/encoders to be used to limit concurrency and peak memory consumption as a result. (to unblock [MM-10166](https://mattermost.atlassian.net/browse/MM-10166))
- Improved memory consumption for transparent PNG uploads. This can be considered a temporary solution until [MM-36111](https://mattermost.atlassian.net/browse/MM-36111) is addressed. I didn't want to commit to such a possibly breaking change so waiting for 6.0 to do it.

Here's a quick test for a sample PNG file with transparency:

![imgfill](https://user-images.githubusercontent.com/1832946/120348460-7deb0200-c2fd-11eb-8896-79a2a4a3b31e.png)

Annotations indicate the following:

1. Sample image upload (on master).
2. Server restart.
3. Sample image upload (on PR changes).

Now that the filling operation is done in-place we save one full allocation, which for a raw image can be considerable (up to ~196MB for our current resolution limit of 24MP).

#### Benchmark

```
name                                                     old time/op    new time/op    delta
FillImageTransparency/10MPx_fully_transparent_RGBA-8       6.70ms ± 4%    2.77ms ± 4%   -58.64%  (p=0.000 n=10+10)
FillImageTransparency/10MPx_partially_opaque_RGBA-8        6.71ms ± 2%    2.48ms ±10%   -63.13%  (p=0.000 n=10+10)
FillImageTransparency/10MPx_fully_opaque_RGBA-8            6.72ms ± 1%    0.71ms ±17%   -89.38%  (p=0.000 n=8+10)
FillImageTransparency/10MPx_fully_opaque_palette-8         17.4ms ± 3%     0.3ms ± 9%   -98.22%  (p=0.000 n=9+10)
FillImageTransparency/10MPx_fully_transparent_palette-8    19.2ms ± 2%     0.3ms ± 5%   -98.32%  (p=0.000 n=9+10)

name                                                     old alloc/op   new alloc/op   delta
FillImageTransparency/10MPx_fully_transparent_RGBA-8       4.01MB ± 0%    0.00MB ± 8%  -100.00%  (p=0.000 n=9+10)
FillImageTransparency/10MPx_partially_opaque_RGBA-8        4.01MB ± 0%    0.00MB ± 4%  -100.00%  (p=0.000 n=10+10)
FillImageTransparency/10MPx_fully_opaque_RGBA-8            4.01MB ± 0%    0.00MB ± 4%  -100.00%  (p=0.000 n=10+10)
FillImageTransparency/10MPx_fully_opaque_palette-8         4.01MB ± 0%    0.00MB ±42%  -100.00%  (p=0.000 n=10+9)
FillImageTransparency/10MPx_fully_transparent_palette-8    4.01MB ± 0%    0.00MB ±38%  -100.00%  (p=0.000 n=9+10)

name                                                     old allocs/op  new allocs/op  delta
FillImageTransparency/10MPx_fully_transparent_RGBA-8         4.00 ± 0%      1.00 ± 0%   -75.00%  (p=0.000 n=10+10)
FillImageTransparency/10MPx_partially_opaque_RGBA-8          4.00 ± 0%      1.00 ± 0%   -75.00%  (p=0.000 n=10+10)
FillImageTransparency/10MPx_fully_opaque_RGBA-8              4.00 ± 0%      1.00 ± 0%   -75.00%  (p=0.000 n=10+10)
FillImageTransparency/10MPx_fully_opaque_palette-8           4.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
FillImageTransparency/10MPx_fully_transparent_palette-8      4.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```

#### Ticket

https://mattermost.atlassian.net/browse/MM-29241

#### Release Note

```release-note
Improved memory performance for large image uploads, particularly PNGs with transparency.
```