@streamer45 2021-06-03
@saturninoabril 

For this I am thinking a smoke-test specifically around image uploads (profile picture, post attachments, brand image) and making sure thumbnail/previews work as expected.